requirements.txt
torch~=2.0.1
transformers~=4.30.2
Pillow~=9.5.0
Flask~=2.3.2